+++ apps/web/src/apis/domains/team/apply/endpoints.ts
