+++ apps/web/src/apis/domains/team/apply/schemas.ts
