+++ apps/web/src/apis/domains/team/apply/types.ts
